run afterbuild plugins in serve as well
