+++ examples/with-dev-source/abell.config.js
@@ -1,0 +1,3 @@
+module.exports = {
+  plugins: ['plugins/dev-articles.js']
+}
