+++ examples/with-dev-source/theme/data/articles.json
@@ -1,0 +1,37 @@
+[
+  {
+    "url": "https://dev.to/saurabhdaware/i-rebuilt-my-portfolio-now-it-has-100-lighthouse-score-here-s-how-i-did-477a",
+    "title": "I rebuilt my portfolio🌻 Now it loads in 1.6s 🎉 Here's how I did ",
+    "tag_list": [
+      "javascript",
+      "webdev",
+      "css",
+      "showdev"
+    ],
+    "positive_reactions_count": 991,
+    "comments_count": 83
+  },
+  {
+    "url": "https://dev.to/saurabhdaware/i-made-dev-to-widget-for-websites-blogs-40p2",
+    "title": "I made DEV.to widget for websites/blogs",
+    "tag_list": [
+      "showdev",
+      "javascript",
+      "webdev"
+    ],
+    "positive_reactions_count": 287,
+    "comments_count": 53
+  },
+  {
+    "url": "https://dev.to/saurabhdaware/i-made-a-text-to-handwriting-tool-to-write-my-college-assignments-for-me-4ko",
+    "title": "I made a Text-to-handwriting tool to write my college assignments for me 🤓",
+    "tag_list": [
+      "showdev",
+      "html",
+      "css",
+      "javascript"
+    ],
+    "positive_reactions_count": 245,
+    "comments_count": 31
+  }
+]
