+++ package.json
@@ -50,6 +50,7 @@
     "eslint-plugin-prettier": "^3.1.3",
     "husky": "^4.2.5",
     "mocha": "^7.1.2",
+    "node-fetch": "^2.6.0",
     "prettier": "^2.0.5",
     "serve": "^11.3.2"
   },
